+++ apps/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ensembleui/react-kitchen-sink
 
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies [dbb6a69]
+  - @ensembleui/react-framework@0.1.41
+  - @ensembleui/react-runtime@0.1.56
+
+## 0.0.47
+
+### Patch Changes
+
+- Updated dependencies [e44293c]
+  - @ensembleui/react-runtime@0.1.56
+
 ## 0.0.46
 
 ### Patch Changes